refactor: reorganize components into the MainPage directory structure
