+++ frontend/src/components/MainPage/Dashboard.tsx
@@ -22,13 +22,7 @@ export default function Dashboard() {
                 <Box sx={{ display: 'flex', gap: '20px' }}>
                     {/* Left Panel - Booking Actions */}
                     <Box sx={{ flex: 1, maxWidth: '30rem' }}>
-                        <BookingPanel
-                            localDate="01/22/2025"
-                            localTime="09:15:42"
-                            timeZone="GMT +01:00 (Berlin)"
-                            bookingType="B1 Arrival"
-                            employeeName="Jane, Patrick (00293)"
-                        />
+                        <BookingPanel />
                     </Box>
 
                     {/* Right Panel - Today's Summary */}
